move data_manager
