+++ .gitignore
@@ -1,2 +1,2 @@
 __pycache__
-tests/pose_optimizer/log/
+tests/tools/log/
