+++ tests/pose_optimizer/tools.py
@@ -2,7 +2,7 @@ import numpy as np
 from scipy.spatial.transform import Rotation as scipy_R
 
 from modules.pose_optimizer.optimizer import Optimizer, Keyframe
-from tests.pose_optimizer.data_manager import DataManager
+from tests.tools.data_manager import DataManager
 
 
 def get_random():
